add XLS library
:new: Library which stands for XOBJ Language String. A standardized
prefixed JSON object for translations. Comes with the following
functions:
XLS(data) - Converts a list into a JSON object and prefixes it with
$XL$. Has a stride of two, the first item is the language (use XLS_*)
and the second is the text.

xparse(user, text) - Checks an XLS object and returns a string in the
language of the agent's SL client, or the default language specified by
XLS_LANG_DEFAULT if not found. If the string isn't an XLS string, it's
returned as is.

xme(text) Shortcut for above to fetch a text for llGetOwner()

Parsing should take place before any llRegionSay etc since those are
limited to 1024 bytes. Can easily be implemented into macros.